removed references to wrangler.
